+++ Sources/SendBirdSDK.xcframework/ios-arm64_armv7/SendBirdSDK.framework/Headers/SBDMessageChangeLogsParams.h
@@ -28,12 +28,16 @@
 /// Determines whether to include the replies of the updated messages in the results.
 /// @note The default value is `NO`.
 /// @since 3.0.181
-@property (atomic) BOOL includeReplies;
+/// @deprecated 3.0.236 (Use `replyType` instead.)
+@property (atomic) BOOL includeReplies
+DEPRECATED_ATTRIBUTE;
 
 /// Determines whether to include the parent message text in the results when the updated messages are threaded replies. If the type of the parent message is [`SBDUserMessage`](../Classes/SBDUserMessage.html), the value is a [`message`](../Classes/SBDUserMessage.html#//api/name/message). If it is [`SBDFileMessage`](../Classes/SBDFileMessage.html), the value is the [`name`](../Classes/SBDFileMessage.html#//api/name/name) of the uploaded file.
 /// @note The default value is `NO`.
 /// @since 3.0.181
-@property (atomic) BOOL includeParentMessageText;
+/// @deprecated 3.0.236 (Use `includeParentMessageInfo` instead.)
+@property (atomic) BOOL includeParentMessageText
+DEPRECATED_ATTRIBUTE;
 
 /// Determines whether to include the thread information of the updated messages in the results when the results contain parent messages.
 /// @note The default value is `NO`.
@@ -45,13 +49,23 @@
 /// @since 3.0.232
 @property (atomic) BOOL includePollDetails;
 
-/// Create `SBDPreviousMessageListQuery` from a given SBDPreviousMessageListQuery instance.
+/// Determines wheter to include information on parent message.
+/// @note The default value is `NO`
+/// @since 3.0.236
+@property (atomic) BOOL includeParentMessageInfo;
+
+/// Determines which reply types to include in the results.
+/// @note The default value is `SBDReplyTypeNone`
+/// @since 3.0.236
+@property (atomic) SBDReplyType replyType;
+
+/// Create `SBDMessageChangeLogsParams` from a given `SBDPreviousMessageListQuery` instance.
 /// @param query previous message list query
 /// @return A created instance of `SBDMessageChangeLogsParams`
 /// @since 3.0.185
 + (nonnull SBDMessageChangeLogsParams *)createWithPreviousMessageListQuery:(nullable SBDPreviousMessageListQuery *)query;
 
-/// Create `SBDMessageListParams` from a given SBDMessageListParams instance.
+/// Create `SBDMessageChangeLogsParams` from a given `SBDMessageListParams` instance.
 /// @param params previous message list query
 /// @return A created instance of `SBDMessageChangeLogsParams`
 /// @since 3.0.185