+++ Sources/SendBirdSDK.xcframework/ios-arm64_i386_x86_64-simulator/SendBirdSDK.framework/Headers/SBDMessageListParams.h
@@ -64,13 +64,15 @@
 
 /// Determines whether to include the replies to the messages in the results.
 /// @note The default value is `NO`.
-/// @since 3.0.181
-@property (atomic) BOOL includeReplies;
+/// @deprecated 3.0.236 (Use `replyType` instead.)
+@property (nonatomic) BOOL includeReplies
+DEPRECATED_ATTRIBUTE;
 
 /// Determines whether to include the parent message text in the results when the messages are replies in a thread. If the type of the parent message is [`SBDUserMessage`](../Classes/SBDUserMessage.html), the value is a [`message`](../Classes/SBDUserMessage.html#//api/name/message). If it is [`SBDFileMessage`](../Classes/SBDFileMessage.html), the value is the [`name`](../Classes/SBDFileMessage.html#//api/name/name) of the uploaded file.
 /// @note The default value is `NO`.
-/// @since 3.0.181
-@property (atomic) BOOL includeParentMessageText;
+/// @deprecated 3.0.236 (Use `includeParentMessageInfo` instead.)
+@property (nonatomic) BOOL includeParentMessageText
+DEPRECATED_ATTRIBUTE;
 
 /// Determines whether to include the thread information of the messages in the results when the results contain root messages.
 /// @note The default value is `NO`.
@@ -81,6 +83,16 @@
 /// @note The default value is `NO`
 /// @since 3.0.232
 @property (atomic) BOOL includePollDetails;
+
+/// Determines whether to include parent message info.
+/// @note The default value is `NO`
+/// @since 3.0.236
+@property (atomic) BOOL includeParentMessageInfo;
+
+/// Determines which reply types to include.
+/// @note The default value is `SBDReplyTypeNone`
+/// @since 3.0.236
+@property (atomic) SBDReplyType replyType;
 
 /// Determines whether to include only messages from the subChannel to which you belong in the results.
 /// @note The default value is `NO`.